requirements.txt
Django==1.11.7
django-haystack==2.6.1
django-redis==4.8.0
django-tinymce==2.6.0
jieba==0.39
MySQL-python==1.2.5
olefile==0.44
Pillow==4.3.0
PyMySQL==0.7.11
pytz==2017.3
redis==2.10.6
Whoosh==2.7.4
